Refactoring to create libweb and move all code there to allow testing
